# Conflicts:
#	Assets/PatchKit Patcher/Scripts/AppUpdater/Commands/InstallContentCommand.cs
#	Assets/PatchKit Patcher/Scripts/AppUpdater/Commands/InstallDiffCommand.cs
#	Assets/PatchKit Patcher/Scripts/AppUpdater/Commands/RepairFilesCommand.cs
#	Assets/PatchKit Patcher/Scripts/PatcherError.cs

+++ Assets/PatchKit Patcher/Scripts/AppUpdater/Commands/InstallDiffCommand.cs
@@ -40,6 +40,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
         private AppContentSummary _contentSummary;
         private AppDiffSummary _diffSummary;
         private Pack1Meta _pack1Meta;
+        private MapHashExtractedFiles _mapHashExtractedFiles;
 
         public InstallDiffCommand([NotNull] string packagePath, string packageMetaPath, string packagePassword,
             int versionId,
@@ -79,6 +80,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
             _localData = localData;
             _localMetaData = localMetaData;
             _remoteMetaData = remoteMetaData;
+            _mapHashExtractedFiles = new MapHashExtractedFiles();
         }
 
         public override void Prepare([NotNull] UpdaterStatus status, CancellationToken cancellationToken)
@@ -304,10 +306,10 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
             {
                 case "zip":
                     usedSuffix = string.Empty;
-                    return new ZipUnarchiver(_packagePath, destinationDir, _packagePassword);
+                    return new ZipUnarchiver(_packagePath, destinationDir, _mapHashExtractedFiles, _packagePassword);
                 case "pack1":
                     usedSuffix = Suffix;
-                    return new Pack1Unarchiver(_packagePath, _pack1Meta, destinationDir, _packagePassword, Suffix);
+                    return new Pack1Unarchiver(_packagePath, _pack1Meta, destinationDir, _mapHashExtractedFiles, _packagePassword, Suffix);
                 default:
                     throw new UnknownPackageCompressionModeException(string.Format("Unknown compression method: {0}",
                         _diffSummary.CompressionMethod));
@@ -442,7 +444,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
                 }
                 else
                 {
-                    AddFile(entryName, packageDirPath, suffix, cancellationToken);
+                    AddFile(entryName, packageDirPath, suffix, cancellationToken, i);
                 }
 
                 _addFilesStatusReporter.Progress.Value = (i + 1) / (double) _diffSummary.AddedFiles.Length;
@@ -471,36 +473,53 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
             _logger.LogDebug("Add directory entry processed.");
         }
 
-        private void AddFile(string fileName, string packageDirPath, string suffix, CancellationToken cancellationToken)
+        private void AddFile(string fileName, string packageDirPath, string suffix, CancellationToken cancellationToken,
+            int fileIndex)
         {
             _logger.LogDebug(string.Format("Processing add file entry {0}", fileName));
 
             var filePath = _localData.Path.PathCombine(fileName);
             _logger.LogTrace("filePath = " + filePath);
-            var sourceFilePath = Path.Combine(packageDirPath, fileName + suffix);
-            _logger.LogTrace("sourceFilePath = " + sourceFilePath);
-
-            if (!File.Exists(sourceFilePath))
+#if UNITY_STANDALONE_WIN
+            if (filePath.Length > 259)
             {
-                throw new MissingFileFromPackageException(string.Format("Cannot find file {0} in diff package.",
-                    fileName));
+                throw new FilePathTooLongException(string.Format("Cannot install file {0}, the destination path length has exceeded Windows path length limit (260).", filePath)); 
             }
+#endif
+            string nameHash;
+            if (_mapHashExtractedFiles.TryGetHash(fileName, out nameHash))
+            {
+                var sourceFilePath = Path.Combine(packageDirPath, nameHash + suffix);
+                _logger.LogTrace("sourceFilePath = " + sourceFilePath);
 
-            _logger.LogDebug("Creating file parent directories in local data...");
-            var fileParentDirPath = Path.GetDirectoryName(filePath);
-            _logger.LogTrace("fileParentDirPath = " + fileParentDirPath);
-            //TODO: Assert that fileParentDirPath is not null
-            // ReSharper disable once AssignNullToNotNullAttribute
-            DirectoryOperations.CreateDirectory(fileParentDirPath, cancellationToken);
-            _logger.LogDebug("File parent directories created in local data.");
+                if (!File.Exists(sourceFilePath))
+                {
+                    throw new MissingFileFromPackageException(string.Format("Cannot find file {0} in diff package.",
+                        fileName));
+                }
+                
+                _logger.LogDebug("Creating file parent directories in local data...");
+                var fileParentDirPath = Path.GetDirectoryName(filePath);
+                _logger.LogTrace("fileParentDirPath = " + fileParentDirPath);
+                //TODO: Assert that fileParentDirPath is not null
+                // ReSharper disable once AssignNullToNotNullAttribute
+                DirectoryOperations.CreateDirectory(fileParentDirPath, cancellationToken);
+                _logger.LogDebug("File parent directories created in local data.");
 
-            _logger.LogDebug("Copying file to local data (overwriting if needed)...");
-            FileOperations.Copy(sourceFilePath, filePath, true, cancellationToken);
-            _logger.LogDebug("File copied to local data.");
+                _logger.LogDebug("Copying file to local data (overwriting if needed)...");
+                FileOperations.Copy(sourceFilePath, filePath, true, cancellationToken);
+                _logger.LogDebug("File copied to local data.");
 
-            _localMetaData.RegisterEntry(fileName, _versionId);
+                _localMetaData.RegisterEntry(fileName, _versionId, 
+                    _contentSummary.Files.First(x => x.Path == fileName).Size, 
+                    fileIndex == _diffSummary.AddedFiles.Length - 1);
 
-            _logger.LogDebug("Add file entry processed.");
+                _logger.LogDebug("Add file entry processed.");
+            }
+            else
+            {
+                throw new InstallerException(string.Format("Cannot find hash for file {0} in mapHash.", fileName));
+            }
         }
 
         private void ProcessModifiedFiles(string packageDirPath, string suffix, TemporaryDirectory tempDiffDir,
@@ -519,7 +538,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
 
                 if (!entryName.EndsWith("/"))
                 {
-                    PatchFile(entryName, packageDirPath, suffix, tempDiffDir, cancellationToken);
+                    PatchFile(entryName, packageDirPath, suffix, tempDiffDir, cancellationToken, i);
                 }
 
                 _modifiedFilesStatusReporter.Progress.Value = (i + 1) / (double) _diffSummary.ModifiedFiles.Length;
@@ -537,7 +556,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
 
         private void PatchFile(
             string fileName, string packageDirPath, string suffix,
-            TemporaryDirectory tempDiffDir, CancellationToken cancellationToken)
+            TemporaryDirectory tempDiffDir, CancellationToken cancellationToken, int fileIndex)
         {
             _logger.LogDebug(string.Format("Processing patch file entry {0}", fileName));
 
@@ -565,36 +584,46 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
             {
                 _logger.LogDebug("Patching is necessary. Generating new file with patched content...");
 
-                var sourceDeltaFilePath = Path.Combine(packageDirPath, fileName + suffix);
-                _logger.LogTrace("sourceDeltaFilePath = " + sourceDeltaFilePath);
-
-                if (!File.Exists(sourceDeltaFilePath))
+                string nameHash;
+                if (_mapHashExtractedFiles.TryGetHash(fileName, out nameHash))
                 {
-                    throw new MissingFileFromPackageException(string.Format(
-                        "Cannot find delta file {0} in diff package.",
-                        fileName));
+                    var sourceDeltaFilePath = Path.Combine(packageDirPath, nameHash + suffix);
+                    _logger.LogTrace("sourceDeltaFilePath = " + sourceDeltaFilePath);
+
+                    if (!File.Exists(sourceDeltaFilePath))
+                    {
+                        throw new MissingFileFromPackageException(string.Format(
+                            "Cannot find delta file {0} in diff package.",
+                            fileName));
+                    }
+
+                    var newFilePath = tempDiffDir.GetUniquePath();
+                    _logger.LogTrace("newFilePath = " + newFilePath);
+
+                    var filePatcher = new FilePatcher(filePath, sourceDeltaFilePath, newFilePath);
+                    filePatcher.Patch();
+
+                    _logger.LogDebug("New file generated. Deleting old file in local data...");
+                    FileOperations.Delete(filePath, cancellationToken);
+
+                    _logger.LogDebug("Old file deleted. Moving new file to local data...");
+                    FileOperations.Move(newFilePath, filePath, cancellationToken);
+
+                    _logger.LogDebug("New file moved.");
                 }
-
-                var newFilePath = tempDiffDir.GetUniquePath();
-                _logger.LogTrace("newFilePath = " + newFilePath);
-
-                var filePatcher = new FilePatcher(filePath, sourceDeltaFilePath, newFilePath);
-                filePatcher.Patch();
-
-                _logger.LogDebug("New file generated. Deleting old file in local data...");
-                FileOperations.Delete(filePath, cancellationToken);
-
-                _logger.LogDebug("Old file deleted. Moving new file to local data...");
-                FileOperations.Move(newFilePath, filePath, cancellationToken);
-
-                _logger.LogDebug("New file moved.");
+                else
+                {
+                    throw new InstallerException(string.Format("Cannot find hash for file {0} in mapHash.", fileName));
+                }
             }
             else
             {
                 _logger.LogDebug("Patching is not necessary. File content is the same as in previous version.");
             }
-
-            _localMetaData.RegisterEntry(fileName, _versionId);
+            
+            _localMetaData.RegisterEntry(fileName, _versionId, 
+                _contentSummary.Files.First(x => x.Path == fileName).Size, 
+                fileIndex == _diffSummary.ModifiedFiles.Length - 1);
 
             _logger.LogDebug("Patch file entry processed.");
         }
